Enable request for validation
Validation requests are enabled. A validation request to the Orcfax
Node is synonymous with asking for a price to be placed on-chain.

If the request fails here, we don't retry as we would retry
connecting to the price monitor endpoint. This might not be the
right approach, we may need separate retry conditions for this
new function.